+++ back-end/ms-client/src/main/java/com/petscare/msclient/ClientServiceApplication.java
@@ -1,0 +1,15 @@
+package com.petscare.msclient;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class ClientServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ClientServiceApplication.class, args);
+    }
+
+
+}
