+++ back-end/ms-client/src/main/java/com/petscare/msclient/repository/IClientRepository.java
@@ -1,0 +1,10 @@
+package com.petscare.msclient.repository;
+
+import com.petscare.msclient.model.Client;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IClientRepository extends JpaRepository<Client, Long> {
+
+}
