+++ back-end/ms-client/src/test/java/com/petscare/msclient/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.petscare.msclient;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
