+++ back-end/ms-consult/src/main/java/com/petscare/msconsult/ConsultServiceApplication.java
@@ -1,0 +1,15 @@
+package com.petscare.msconsult;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class ConsultServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ConsultServiceApplication.class, args);
+    }
+
+
+}
