+++ back-end/ms-consult/src/main/java/com/petscare/msconsult/model/Consult.java
@@ -1,0 +1,30 @@
+package com.petscare.msconsult.model;
+
+import com.fasterxml.jackson.annotation.JsonIgnore;
+import jakarta.persistence.*;
+import lombok.Data;
+
+import java.time.LocalDate;
+
+@Data
+@Entity
+@Table(name = "Consult")
+public class Consult {
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    @Column(name = "created_at")
+    private LocalDate cratedAt;
+
+    @Column(name = "diagnostic")
+    private String diagnostic;
+
+    @Column(name = "drugs")
+    private String drugs;
+
+    @Column(name = "exams")
+    private String exams;
+
+
+}
