+++ back-end/ms-consult/src/test/java/com/petscare/msconsult/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.petscare.msconsult;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
