+++ back-end/ms-pet/src/main/java/com/petscare/mspet/PetServiceApplication.java
@@ -1,0 +1,15 @@
+package com.petscare.mspet;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+
+@SpringBootApplication
+public class PetServiceApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(PetServiceApplication.class, args);
+    }
+
+
+}
