+++ back-end/ms-pet/src/main/java/com/petscare/mspet/repository/IPetRepository.java
@@ -1,0 +1,10 @@
+package com.petscare.mspet.repository;
+
+import com.petscare.mspet.model.Pet;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IPetRepository extends JpaRepository<Pet, Long> {
+
+}
