+++ back-end/ms-pet/src/main/java/com/petscare/mspet/repository/IPetTypeRepository.java
@@ -1,0 +1,9 @@
+package com.petscare.mspet.repository;
+
+import com.petscare.mspet.model.PetType;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
+
+@Repository
+public interface IPetTypeRepository  extends JpaRepository<PetType, Long> {
+}
