+++ back-end/ms-pet/src/test/java/com/petscare/mspet/UserServiceApplicationTests.java
@@ -1,0 +1,13 @@
+package com.petscare.mspet;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class UserServiceApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
