--- back-end/ms-user/src/main/java/com/petscare/msuser/UserServiceApplication.java
@@ -1,15 +1,0 @@
-package com.petscare.msuser;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-
-@SpringBootApplication
-public class UserServiceApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(UserServiceApplication.class, args);
-    }
-
-
-}
