--- back-end/ms-user/src/main/java/com/petscare/msuser/repository/IPetRepository.java
@@ -1,10 +1,0 @@
-package com.petscare.msuser.repository;
-
-import com.petscare.msuser.model.Pet;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface IPetRepository extends JpaRepository<Pet, Long> {
-
-}
