--- back-end/ms-user/src/main/java/com/petscare/msuser/repository/IUserRepository.java
@@ -1,10 +1,0 @@
-package com.petscare.msuser.repository;
-
-import com.petscare.msuser.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-@Repository
-public interface IUserRepository extends JpaRepository<User, Long> {
-
-}
