--- back-end/ms-user/src/main/java/com/petscare/msuser/service/IPetService.java
@@ -1,31 +1,0 @@
-package com.petscare.msuser.service;
-
-import com.petscare.msuser.model.Pet;
-import com.petscare.msuser.model.PetClinicalHistory;
-import com.petscare.msuser.model.PetType;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface IPetService {
-    //Pets
-    List<Pet> getAll();
-    Optional<Pet> getPetById(Long id);
-    Pet createPet(Pet pet);
-    void deletePet(Long id) throws Exception;
-    Pet updatePet(Pet pet) throws Exception;
-
-    //Types
-    List<PetType> getAllType();
-    Optional<PetType> getPetTypeById(Long id);
-    PetType createPetType(PetType petType);
-    void deletePetType(Long id) throws Exception;
-    PetType updatePetType(PetType petType) throws Exception;
-
-    //History
-    List<PetClinicalHistory> getAllClinicalHistory();
-    Optional<PetClinicalHistory> getPetClinicalHistoryById(Long id);
-    PetClinicalHistory createPetClinicalHistory(PetClinicalHistory petClinicalHistory);
-    void deletePetClinicalHistory(Long id) throws Exception;
-    PetClinicalHistory updatePetClinicalHistory(PetClinicalHistory petClinicalHistory) throws Exception;
-}
