--- back-end/ms-user/src/test/java/com/petscare/msuser/UserServiceApplicationTests.java
@@ -1,13 +1,0 @@
-package com.petscare.msuser;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class UserServiceApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
